Remove token.go utils file
